Refactor experiment creation for group A
