+++ Assets/Scripts/Experiments/Patrolling/GroupA/GroupAExperimentHelpers.cs
@@ -1,0 +1,84 @@
+// Copyright 2025 MAEPS
+// 
+// This file is part of MAEPS
+// 
+// MAEPS is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the
+// Free Software Foundation, either version 3 of the License, or (at your option)
+// any later version.
+// 
+// MAEPS is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+// or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+// Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License along
+// with MAEPS. If not, see http://www.gnu.org/licenses/.
+// 
+// Contributors: 
+// Henrik van Peet,
+// Mads Beyer Mogensen,
+// Puvikaran Santhirasegaram
+
+using System.Collections.Generic;
+
+using Maes.Algorithms.Patrolling;
+using Maes.Map.Generators;
+using Maes.Simulation.Patrolling;
+
+namespace Maes.Experiments.Patrolling
+{
+    using static Maes.Map.RobotSpawners.RobotSpawner<IPatrollingAlgorithm>;
+
+    using MySimulationScenario = PatrollingSimulationScenario;
+
+    internal static class GroupAExperimentHelpers
+    {
+        public static MySimulationScenario ScenarioConstructor(int seed, string algorithmName, CreateAlgorithmDelegate algorithm, BuildingMapConfig mapConfig, int robotCount)
+        {
+            return new MySimulationScenario(
+                                        seed: seed,
+                                        totalCycles: GroupAParameters.AmountOfCycles,
+                                        stopAfterDiff: false,
+                                        robotSpawner: (buildingConfig, spawner) => spawner.SpawnRobotsApart(
+                                            collisionMap: buildingConfig,
+                                            seed: seed,
+                                            numberOfRobots: robotCount,
+                                            createAlgorithmDelegate: algorithm),
+                                        mapSpawner: generator => generator.GenerateMap(mapConfig),
+                                        robotConstraints: GroupAParameters.StandardRobotConstraints,
+                                        statisticsFileName:
+                                        $"{algorithmName}-seed-{seed}-size-{mapConfig.HeightInTiles}-robots-{robotCount}-constraints-{GroupAParameters.StandardRobotConstraintName}-BuldingMap-SpawnApart");
+        }
+
+        public static MySimulationScenario ScenarioConstructor(int seed, string algorithmName, CreateAlgorithmDelegate algorithm, CaveMapConfig mapConfig, int robotCount)
+        {
+            return new MySimulationScenario(
+                                       seed: seed,
+                                       totalCycles: GroupAParameters.AmountOfCycles,
+                                       stopAfterDiff: false,
+                                       robotSpawner: (buildingConfig, spawner) => spawner.SpawnRobotsApart(
+                                           collisionMap: buildingConfig,
+                                           seed: seed,
+                                           numberOfRobots: robotCount,
+                                           createAlgorithmDelegate: algorithm),
+                                       mapSpawner: generator => generator.GenerateMap(mapConfig),
+                                       robotConstraints: GroupAParameters.StandardRobotConstraints,
+                                       statisticsFileName:
+                                       $"{algorithmName}-seed-{seed}-size-{mapConfig.HeightInTiles}-robots-{robotCount}-constraints-{GroupAParameters.StandardRobotConstraintName}-BuldingMap-SpawnApart");
+        }
+
+        public static IEnumerable<MySimulationScenario> CreateScenarios(int seed, string algorithmName, CreateAlgorithmDelegate algorithm, int robotCount = GroupAParameters.StandardRobotCount, int mapSize = GroupAParameters.StandardMapSize)
+        {
+            var scenarios = new List<MySimulationScenario>();
+            var buildingMapConfig = new BuildingMapConfig(seed, widthInTiles: mapSize, heightInTiles: mapSize, brokenCollisionMap: false);
+            var caveMapConfig = new CaveMapConfig(seed, widthInTiles: mapSize, heightInTiles: mapSize, brokenCollisionMap: false);
+            var scenarioBuilding = ScenarioConstructor(seed, algorithmName, algorithm, buildingMapConfig, robotCount);
+            var scenarioCave = ScenarioConstructor(seed, algorithmName, algorithm, caveMapConfig, robotCount);
+            scenarios.Add(scenarioBuilding);
+            scenarios.Add(scenarioCave);
+            return scenarios;
+        }
+
+    }
+}
